DAOS-15084 tests: move insert_objects copy_remote_to_local into ddb
These functions are only used for ddb tests.
